add the start of a domains track
