+++ packages/jbrowse-web/src/util/functionStrings.js
@@ -13,25 +13,30 @@ export function stringToFunction(str, options = {}) {
 
   const cacheKey = `${verifyFunctionSignature &&
     verifyFunctionSignature.join(',')}|${str}`
-  if (compilationCache[cacheKey]) return compilationCache[cacheKey]
+  if (!compilationCache[cacheKey]) {
+    const match = functionRegexp.exec(str)
+    if (!match)
+      throw new Error('string does not appear to be a function declaration')
+    const paramList = match[1].split(',').map(s => s.trim())
+    let code = match[2].replace(/}\s*$/, '')
+    if (verifyFunctionSignature) {
+      // check number of arguments passed by calling code at runtime.
+      // NOTE: we don't check the number of arguments in the callback code itself,
+      // callback authors are free to ignore the arguments if they want
 
-  const match = functionRegexp.exec(str)
-  if (!match)
-    throw new Error('string does not appear to be a function declaration')
-  const paramList = match[1].split(',').map(s => s.trim())
-  let code = match[2].replace(/}\s*$/, '')
-  if (verifyFunctionSignature) {
-    // check number of arguments passed by calling code at runtime.
-    // NOTE: we don't check the number of arguments in the callback code itself,
-    // callback authors are free to ignore the arguments if they want
-
-    code = `if (arguments.length !== ${
-      verifyFunctionSignature.length
-    }) throw new Error("incorrect number of arguments provided to callback.  function signature is (${verifyFunctionSignature.join(
-      ', ',
-    )}) but "+arguments.length+" arguments were passed");\n${code}`
+      code = `if (arguments.length !== ${
+        verifyFunctionSignature.length
+      }) throw new Error("incorrect number of arguments provided to callback.  function signature is (${verifyFunctionSignature.join(
+        ', ',
+      )}) but "+arguments.length+" arguments were passed");\n${code}`
+    }
+    const compiled = new Function(...paramList, `"use strict"; ${code}`) // eslint-disable-line no-new-func
+    compilationCache[cacheKey] = compiled
   }
-  const func = new Function(...paramList, `"use strict"; ${code}`) // eslint-disable-line
-  compilationCache[cacheKey] = func
+
+  let func = compilationCache[cacheKey]
+  if (options.bind) {
+    func = func.bind(...options.bind)
+  }
   return func
 }
