+++ packages/jbrowse-web/src/util/range.js
@@ -30,4 +30,8 @@ export function intersection2(left1, right1, left2, right2) {
   return undefined
 }
 
+export function doesIntersect2(left1, right1, left2, right2) {
+  return right1 > left2 && left1 < right2
+}
+
 export function removeme() {} // remove after we have more functions in here
